Login Page

+++ app/Http/Controllers/NoteController.php
@@ -14,6 +14,7 @@ class NoteController extends Controller
     public function index()
     {
         $notes = Note::query()
+            ->where('user_id', request()->user()->id)
             ->orderBy('created_at', 'desc')
             ->paginate(6);
 
@@ -26,6 +27,10 @@ class NoteController extends Controller
      */
     public function create(Note $note)
     {
+        if ($note->user_id != auth()->user()->id) {
+            abort(403);
+        }
+
         return view("note.create", ['note' => $note]);
     }
 
@@ -37,7 +42,7 @@ class NoteController extends Controller
         $data = $request->validate([
             'note' => ['required', 'string'],
         ]);
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
         return to_route('note.index', $note)->with('messages', 'Note created successfully');
@@ -48,6 +53,10 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id != auth()->user()->id) {
+            abort(403);
+        }
+
         return view("note.show", ['note' => $note]);
     }
 
@@ -56,6 +65,10 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if ($note->user_id != auth()->user()->id) {
+            abort(403);
+        }
+
         return view("note.edit", ['note' => $note]);
     }
 
@@ -64,6 +77,10 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id != auth()->user()->id) {
+            abort(403);
+        }
+
         $data = $request->validate([
             'note' => ['required', 'string'],
         ]);
@@ -77,6 +94,10 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id != auth()->user()->id) {
+            abort(403);
+        }
+
         $note->delete();
         return to_route('note.index', $note)->with('messages', 'Note Delete successfully');
     }
